extract emoji board tabs component
