+++ src/app/components/emoji-board/components/index.tsx
@@ -1,0 +1,2 @@
+export * from './SearchInput';
+export * from './Tabs';
