+++ src/app/components/emoji-board/index.ts
@@ -1,1 +1,2 @@
 export * from './EmojiBoard';
+export * from './types';
